Fundamentally changed how the laser works. Previously it would take 500mb out of the tank, infuse that and then reinsert that liquid into the tank (causing it to mix with what is already there). This method has some issues. Say you have a laser catalyst that adds 10% purity with a maximum of 100%. If the tank contains 1000mb of RCL at 95% purity then with the previous system it would extract 500mb (which is still at 95% purity) add 10% purity which will end up at 100% purity since it is capped to 100 and then reinsert that 500mb of 100% pure liquid into the tank. This would result in 1000 mb RCL of 97.5% purity. This is wasteful since essentially the catalyst was only half used.
With the new system it will always operate on the entire tank at once. So with a tank of 1000mb of RCL at 95% purity it will get all that liquid, add 5% purity (since the catalyst adds 10% for 500mb it would add 5% for 1000mb) and reinsert 1000 mb of 100% pure liquid. This is much more correct and less wasteful.

When purity goes below 0% the RCL that is destroyed is always 200mb now (per consumed catalyst).